inference/requirements.txt
tensorflow==2.15.0
numpy==1.26.4
pillow==10.4.0
fastapi==0.111.0
uvicorn[standard]==0.30.0
